+++ public/scripts.js
@@ -49,14 +49,10 @@ document.addEventListener('DOMContentLoaded', () => {
 
   function hideAppBar() {
     const mouseIsInsideWindow = companion.insideWindow();
-    const mouseIsIdle = companion.mouseIsIdle();
     const isHidden = appBar.classList.contains('app-bar--hidden');
     const isFullscreen = companion.isFullscreen();
 
-    if (
-      (!mouseIsInsideWindow && !isHidden) ||
-      (mouseIsIdle && !isHidden && isFullscreen)
-    ) {
+    if ((!mouseIsInsideWindow && !isHidden) || isFullscreen) {
       appBar.classList.add('app-bar--hidden');
       webviewWrapper.classList.add('webview--full');
     }
@@ -66,8 +62,9 @@ document.addEventListener('DOMContentLoaded', () => {
     const mouseIsInsideWindow = companion.insideWindow();
     const mouseIsIdle = companion.mouseIsIdle();
     const isHidden = appBar.classList.contains('app-bar--hidden');
+    const isFullscreen = companion.isFullscreen();
 
-    if (mouseIsInsideWindow && isHidden && !mouseIsIdle) {
+    if ((mouseIsInsideWindow && isHidden && !mouseIsIdle) && !isFullscreen) {
       appBar.classList.remove('app-bar--hidden');
       webviewWrapper.classList.remove('webview--full');
     }
@@ -106,7 +103,7 @@ document.addEventListener('DOMContentLoaded', () => {
   window.setInterval(checkHistoryButtons, 1000);
   window.setInterval(updateURL, 1000);
   window.setInterval(showAppBar, 1000);
-  window.setInterval(hideAppBar, 2000);
+  window.setInterval(hideAppBar, 3000);
 
   searchBar.addEventListener('submit', (e) => {
     e.preventDefault();